Type: Feat
Body:
투자 금액에 따른 납부비율 선정하는 서비스 및 컨트롤러

Footer:

Related to: #7

+++ src/main/java/com/shinhan/entity/InvestmentEntity.java
@@ -43,7 +43,7 @@ public class InvestmentEntity {
 
     private Boolean isActive; // 활성화 여부
 
-    private Integer refundRate; // 환급 비율 (%)
+    private Double refundRate; // 환급 비율 (%)
 
     private Integer maxInvestment; // 최대 투자 금액
